fix(qwen-web): split oversized specialty test file + fix short-id fixtures + add version header test
Three follow-ups to keep this PR release-green:

1. provider-validation-specialty.test.ts exceeded the frozen file-size
   cap (config/quality/file-size-baseline.json). Split the qwen-web
   validator tests into a new tests/unit/provider-validation-qwen-web-5855.test.ts
   so the shared god-file stays under its cap.

2. The mock user fixtures for the qwen-web /api/v1/auths/ probe used
   `id: "u-1"`, which is shorter than the 8-char minimum enforced by
   the tightened top-level user-id check added in this PR — fixed to
   `"u-1234567"` in both the split test and qwen-web-cookie-validation-3958.test.ts.

3. Added an explicit unit test asserting the executor sends the
   `version: 0.2.66` SPA header on the v2 chat/completions request.
